fixed pull #1266 , optimizer code
