+++ deployer/src/main/resources/example/instance.properties
@@ -48,5 +48,5 @@ canal.mq.topic=example
 canal.mq.partition=0
 # hash partition config
 #canal.mq.partitionsNum=3
-#canal.mq.partitionHash=.*\\..*.$pk$
+#canal.mq.partitionHash=test.table#id^name,.*\\..*
 #################################################
