+++ filter/src/main/java/com/alibaba/otter/canal/filter/aviater/AviaterRegexFilter.java
@@ -129,4 +129,5 @@ public class AviaterRegexFilter implements CanalEventFilter<String> {
     public String toString() {
         return pattern;
     }
+
 }
